Cleanup front-end, more components
