+++ package.json
@@ -25,7 +25,6 @@
     },
     "scripts": {
         "test": "jest",
-        "parse-logs": "ts-node parse-logs.ts",
         "postinstall": "cd web && npm install",
         "build-client": "cd web && npm run build",
         "start-server": "ts-node -r dotenv/config index.ts",
