+++ servers/web-socket.ts
@@ -1,10 +1,11 @@
 const { Server } = require('ws');
 
+const MAX_MESSAGE_LENGTH = 1000;
 let sockets = [];
 let messages = [];
 
 export const startWebSocketServer = (server: any) => {
-    const wss = new Server({server});
+    const wss = new Server({ server });
     wss.on("connection", (socket) => {
         sockets.push(socket);
         socket.send(JSON.stringify(messages));
@@ -17,10 +18,14 @@ export const startWebSocketServer = (server: any) => {
 
 const addMessageToQueue = (object: object) => {
     messages.push(object);
-    if (messages.length > 1000) {
+    if (messages.length > MAX_MESSAGE_LENGTH) {
         messages.shift();
     }
 };
+
+export const getMessages = (): any[] => {
+    return messages;
+}
 
 export const emitMessage = (object: object) => {
     let copiedObject: object = {};
