+++ web/package.json
@@ -12,7 +12,6 @@
     "@vitejs/plugin-react-refresh": "^1.3.1",
     "classnames": "^2.3.1",
     "d3": "^7.0.1",
-    "d3-regression": "^1.3.9",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
     "react-tabs": "^3.2.2",
