--- web/src/App.scss
@@ -1,159 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #0c0d16;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-button {
-  font-size: calc(10px + 2vmin);
-}
-
-.display-1,
-.display-2,
-.display-3,
-.display-4 {
-  font-family: "Hack", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"; }
-  .text-white .display-1, .display-1.text-white, .text-white
-  .display-2,
-  .display-2.text-white, .text-white
-  .display-3,
-  .display-3.text-white, .text-white
-  .display-4,
-  .display-4.text-white {
-    text-shadow: -1px -1px rgba(255, 255, 255, 0.2); 
-}
-
-.display-1 {
-  font-size: 4.5rem;
-  font-weight: 300;
-  line-height: 1.2; 
-}
-
-.display-2 {
-  font-size: 3.5rem;
-  font-weight: 300;
-  line-height: 1.2; 
-}
-
-.display-3 {
-  font-size: 2.5rem;
-  font-weight: 300;
-  line-height: 1.2; 
-}
-
-.display-4 {
-  font-size: 2rem;
-  font-weight: 300;
-  line-height: 1.2; 
-}
-
-.vim-caret {
-  -webkit-animation: vimCaret 1s linear infinite;
-  -o-animation: vimCaret 1s linear infinite;
-  animation: vimCaret 1s linear infinite; 
-}
-
-@-webkit-keyframes vimCaret {
-  0% {
-    background-color: transparent; }
-  49% {
-    background-color: transparent; }
-  50% {
-    background-color: rgba(255, 255, 255, 0.2); }
-  100% {
-    background-color: rgba(255, 255, 255, 0.2); } 
-}
-
-@-o-keyframes vimCaret {
-  0% {
-    background-color: transparent; }
-  49% {
-    background-color: transparent; }
-  50% {
-    background-color: rgba(255, 255, 255, 0.2); }
-  100% {
-    background-color: rgba(255, 255, 255, 0.2); } 
-}
-
-@keyframes vimCaret {
-  0% {
-    background-color: transparent; }
-  49% {
-    background-color: transparent; }
-  50% {
-    background-color: rgba(255, 255, 255, 0.2); }
-  100% {
-    background-color: rgba(255, 255, 255, 0.2); } 
-}
-
-.text-mono {
-  font-family: "Hack", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"; 
-}
-
-.grid-container {
-  display: grid;
-  grid-template-columns: auto auto auto auto auto auto;
-  /* background-color: #2196F3; */
-  /* padding: 10px; */
-}
-
-.grid-item {
-  /* background-color: rgba(255, 255, 255, 0.8); */
-  border: 1px solid rgba(255, 255, 255, 0.2);
-  padding: 16px;
-  font-size: 16px;
-  text-align: center;
-}
-
-.link-button {
-  font-size: 16px;
-  background: none!important;
-  border: none;
-  padding: 0!important;
-  font-family: arial, sans-serif;
-  color: #509fb5;
-  text-decoration: underline;
-  cursor: pointer;
-}
-
-.link-button:hover {
-  color: #61dafb;
-}
-
-.react-tabs__tab {
-  padding-left: 36px;
-  padding-right: 36px;
-}
