+++ web/src/components/button/Button.module.scss
@@ -1,0 +1,169 @@
+.btn {
+    display: inline-block;
+    font-weight: 400;
+    color: #e4e2ff;
+    text-align: center;
+    vertical-align: middle;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+    background-color: transparent;
+    border: 1px solid transparent;
+    padding: 0.375rem 0.75rem;
+    font-size: 1rem;
+    line-height: 1.5;
+    border-radius: 6px;
+    -webkit-transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out, border-color 0.15s ease-in-out, -webkit-box-shadow 0.15s ease-in-out;
+    transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out, border-color 0.15s ease-in-out, -webkit-box-shadow 0.15s ease-in-out;
+    -o-transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out, border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
+    transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out, border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
+    transition: color 0.15s ease-in-out, background-color 0.15s ease-in-out, border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out, -webkit-box-shadow 0.15s ease-in-out;
+}
+@media (prefers-reduced-motion: reduce) {
+    .btn {
+        -webkit-transition: none;
+        -o-transition: none;
+        transition: none;
+    }
+}
+.btn:hover {
+    cursor: pointer;
+    color: #e4e2ff;
+    text-decoration: none;
+}
+.btn:focus,
+.btn.focus {
+    outline: 0;
+    -webkit-box-shadow: 0 0 0 0.2rem rgba(113, 76, 223, 0.25);
+    box-shadow: 0 0 0 0.2rem rgba(113, 76, 223, 0.25);
+}
+.btn.disabled,
+.btn:disabled {
+    opacity: 0.65;
+}
+a.btn.disabled,
+fieldset:disabled a.btn {
+    pointer-events: none;
+}
+.btnPrimary {
+    color: #fff;
+    background-color: #714cdf;
+    border-color: #714cdf;
+}
+.btnPrimary:hover {
+    color: #fff;
+    background-color: #572cd9;
+    border-color: #5126d2;
+}
+.btnPrimary:focus,
+.btnPrimary.focus {
+    -webkit-box-shadow: 0 0 0 0.2rem rgba(134, 103, 228, 0.5);
+    box-shadow: 0 0 0 0.2rem rgba(134, 103, 228, 0.5);
+}
+.btnPrimary.disabled,
+.btnPrimary:disabled {
+    color: #fff;
+    background-color: #714cdf;
+    border-color: #714cdf;
+}
+.btnPrimary:not(:disabled):not(.disabled):active,
+.btnPrimary:not(:disabled):not(.disabled).active,
+.show > .btnPrimary.dropdown-toggle {
+    color: #fff;
+    background-color: #5126d2;
+    border-color: #4d24c8;
+}
+.btnPrimary:not(:disabled):not(.disabled):active:focus,
+.btnPrimary:not(:disabled):not(.disabled).active:focus,
+.show > .btnPrimary.dropdown-toggle:focus {
+    -webkit-box-shadow: 0 0 0 0.2rem rgba(134, 103, 228, 0.5);
+    box-shadow: 0 0 0 0.2rem rgba(134, 103, 228, 0.5);
+}
+.btnOutlinePrimary {
+    color: #714cdf;
+    border-color: #714cdf;
+}
+.btnOutlinePrimary:hover {
+    color: #fff;
+    background-color: #714cdf;
+    border-color: #714cdf;
+}
+.btnOutlinePrimary:focus,
+.btnOutlinePrimary.focus {
+    -webkit-box-shadow: 0 0 0 0.2rem rgba(113, 76, 223, 0.5);
+    box-shadow: 0 0 0 0.2rem rgba(113, 76, 223, 0.5);
+}
+.btnOutlinePrimary.disabled,
+.btnOutlinePrimary:disabled {
+    color: #714cdf;
+    background-color: transparent;
+}
+.btnOutlinePrimary:not(:disabled):not(.disabled):active,
+.btnOutlinePrimary:not(:disabled):not(.disabled).active,
+.show > .btnOutlinePrimary.dropdown-toggle {
+    color: #fff;
+    background-color: #714cdf;
+    border-color: #714cdf;
+}
+.btnOutlinePrimary:not(:disabled):not(.disabled):active:focus,
+.btnOutlinePrimary:not(:disabled):not(.disabled).active:focus,
+.show > .btnOutlinePrimary.dropdown-toggle:focus {
+    -webkit-box-shadow: 0 0 0 0.2rem rgba(113, 76, 223, 0.5);
+    box-shadow: 0 0 0 0.2rem rgba(113, 76, 223, 0.5);
+}
+.btnLg,
+.btnGroupLg > .btn {
+    padding: 0.5rem 1rem;
+    font-size: 1.25rem;
+    line-height: 1.5;
+    border-radius: 6px;
+}
+.btnSm,
+.btnGroupSm > .btn {
+    padding: 0.25rem 0.5rem;
+    font-size: 0.875rem;
+    line-height: 1.5;
+    border-radius: 0.2rem;
+}
+.btnShadow {
+    margin-left: 8px;
+}
+.btnShadow.btnPrimary,
+.btnShadow.btnOutlinePrimary {
+    border-color: #986edf;
+    -webkit-transition: none;
+    -o-transition: none;
+    transition: none;
+}
+.btnShadow.btnPrimary,
+.btnShadow.btnPrimary.focus,
+.btnShadow.btnPrimary:focus,
+.btnShadow.btnOutlinePrimary,
+.btnShadow.btnOutlinePrimary.focus,
+.btnShadow.btnOutlinePrimary:focus {
+    -webkit-box-shadow: 3px 3px 0 #986edf, 2px 2px 0 #986edf, 1px 1px 0 #986edf;
+    box-shadow: 3px 3px 0 #986edf, 2px 2px 0 #986edf, 1px 1px 0 #986edf;
+}
+.btnShadow.btnPrimary:hover,
+.btnShadow.btnOutlinePrimary:hover {
+    -webkit-box-shadow: 5px 5px 0 #986edf, 4px 4px 0 #986edf, 3px 3px 0 #986edf, 2px 2px 0 #986edf, 1px 1px 0 #986edf;
+    box-shadow: 5px 5px 0 #986edf, 4px 4px 0 #986edf, 3px 3px 0 #986edf, 2px 2px 0 #986edf, 1px 1px 0 #986edf;
+    -webkit-transform: translate(-2px, -2px);
+    -o-transform: translate(-2px, -2px);
+    transform: translate(-2px, -2px);
+    -webkit-transition: all 0.3s ease;
+    -o-transition: all 0.3s ease;
+    transition: all 0.3s ease;
+}
+.btnShadow.btnPrimary:active,
+.btnShadow.btnOutlinePrimary:active {
+    -webkit-box-shadow: none;
+    box-shadow: none;
+    -webkit-transform: translate(4px, 4px) !important;
+    -o-transform: translate(4px, 4px) !important;
+    transform: translate(4px, 4px) !important;
+    -webkit-transition: all 0.1s ease;
+    -o-transition: all 0.1s ease;
+    transition: all 0.1s ease;
+}
