+++ web/src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Button } from './button/Button';
+export { default as Card } from './card/Card';
+export { default as MultilineChart } from './chart/MultilineChart';
+export { default as LayoutSingleColumn } from './layout/LayoutSingleColumn';
