+++ web/src/pages/ProductPage.scss
@@ -1,14 +1,3 @@
-.react-tabs__tab {
-    font-size: 16px;
-    color: #C8CACC;
-}
-
-.react-tabs__tab.react-tabs__tab--selected {
-    color: #2A343D;
-    background-color: #C8CACC;
-    font-weight: bold;
-}
-
 p.price-data {
     margin-top: 0;
     font-size: 14px;
