+++ web/src/styles/_link.scss
@@ -1,0 +1,20 @@
+a {
+    color: $blue;
+    text-decoration: none;
+    background-color: transparent; 
+}
+a:hover {
+    color: $blue-light;
+    text-decoration: underline; 
+}
+a:not([href]):not([tabindex]) {
+    color: inherit;
+    text-decoration: none; 
+}
+a:not([href]):not([tabindex]):hover, a:not([href]):not([tabindex]):focus {
+    color: inherit;
+    text-decoration: none; 
+}
+a:not([href]):not([tabindex]):focus {
+    outline: 0; 
+}
