+++ web/src/styles/_typography.scss
@@ -1,0 +1,79 @@
+.fine-print {
+    font-size: 12px;
+}
+.display-1, .display-2, .display-3, .display-4, .display-5, .display-6 {
+    font-family: "Hack", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+    font-weight: 300;
+    line-height: 1.2;
+}
+.text-white .display-1, .display-1.text-white, .text-white .display-2, .display-2.text-white, .text-white .display-3, .display-3.text-white, .text-white .display-4, .display-4.text-white {
+    text-shadow: -1px -1px rgba(255, 255, 255, 0.2);
+}
+.display-1 {
+    font-size: 4.5rem;
+}
+.display-2 {
+    font-size: 3.5rem;
+}
+.display-3 {
+    font-size: 2.5rem;
+}
+.display-4 {
+    font-size: 2rem;
+}
+.display-5 {
+    font-size: 1.5rem;
+}
+.display-6 {
+    font-size: 1.25rem;
+}
+.vim-caret {
+    -webkit-animation: vimCaret 1s linear infinite;
+    -o-animation: vimCaret 1s linear infinite;
+    animation: vimCaret 1s linear infinite;
+}
+@-webkit-keyframes vimCaret {
+    0% {
+        background-color: transparent;
+    }
+        49% {
+            background-color: transparent;
+    }
+        50% {
+            background-color: rgba(255, 255, 255, 0.2);
+    }
+        100% {
+            background-color: rgba(255, 255, 255, 0.2);
+    }
+}
+@-o-keyframes vimCaret {
+    0% {
+        background-color: transparent;
+    }
+        49% {
+            background-color: transparent;
+    }
+        50% {
+            background-color: rgba(255, 255, 255, 0.2);
+    }
+        100% {
+            background-color: rgba(255, 255, 255, 0.2);
+    }
+}
+@keyframes vimCaret {
+    0% {
+        background-color: transparent;
+    }
+    49% {
+        background-color: transparent;
+    }
+    50% {
+        background-color: rgba(255, 255, 255, 0.2);
+    }
+    100% {
+        background-color: rgba(255, 255, 255, 0.2);
+    }
+}
+.text-mono {
+    font-family: "Hack", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+}
