+++ web/src/styles/colors/_index.scss
@@ -1,0 +1,2 @@
+@import './variables';
+@import './utilities';
