+++ web/src/styles/colors/_variables.scss
@@ -1,0 +1,40 @@
+$blue: #007bff;
+$blue-light: #16a4de;
+$indigo: #6610f2;
+$purple: #6f42c1;
+$pink: #e83e8c;
+$red: #dc3545;
+$orange: #fd7e14;
+$yellow: #ffc107;
+$green: #28a745;
+$teal: #20c997;
+$cyan: #17a2b8;
+$white: #fff;
+$gray: #1d1e2f;
+$gray-dark: #11111d;
+$light: #dbebfb;
+$dark: #32334a;
+
+$colors: (
+    'blue': $blue,
+    'blue-light': $blue-light,
+    'indigo': $indigo,
+    'purple': $purple,
+    'pink': $pink,
+    'red': $red,
+    'orange': $orange,
+    'yellow': $yellow,
+    'green': $green,
+    'teal': $teal,
+    'cyan': $cyan,
+    'white': $white,
+    'gray': $gray,
+    'gray-dark': $gray-dark,
+    'light': $light,
+    'dark': $dark
+) !default;
+
+$color-properties: (
+    'color': 'color',
+    'bg-color': 'background-color',
+) !default;
