+++ web/src/styles/index.scss
@@ -1,0 +1,10 @@
+//vendor
+@import "react-tabs/style/react-tabs.css";
+//global
+@import './colors/index';
+@import './spacing/index';
+//components
+@import './link';
+@import './typography';
+//overrides
+@import './overrides';
